Organize code well to fix integration tests
