+++ extensions/azurecore/package.json
@@ -228,9 +228,8 @@
               }
             }
           }
-
-          }
         }
+      }
     ],
     "account-type": [
       {
