+++ extensions/azurecore/src/account-provider/interfaces.ts
@@ -5,6 +5,21 @@
 
 import * as azurecore from 'azurecore';
 
+export const enum SettingIds {
+	marm = 'marm',
+	graph = 'graph',
+	msgraph = 'msgraph',
+	arm = 'arm',
+	sql = 'sql',
+	ossrdbms = 'ossrdbms',
+	vault = 'vault',
+	ado = 'ado',
+	ala = 'ala',
+	storage = 'storage',
+	kusto = 'kusto',
+	powerbi = 'powerbi'
+}
+
 /**
  * Mapping of configuration key with the metadata to instantiate the account provider
  */
@@ -18,6 +33,38 @@ export interface ProviderSettings {
 	 * Metadata for the provider
 	 */
 	metadata: azurecore.AzureAccountProviderMetadata;
+}
+
+/**
+ * Custom Provider settings mapping
+ */
+export type ProviderSettingsJson = {
+	name: string,
+	settings: {
+		configKey: string,
+		metadata: {
+			displayName: string,
+			id: string,
+			endpoints: {
+				host: string,
+				microsoftResource: string,
+				graphResource: string,
+				msGraphResource: string,
+				armResource: string,
+				sqlResource: string,
+				azureKeyVaultResource: string,
+				azureLogAnalyticsResource: string,
+				azureStorageResource: {
+					endpoint: string,
+					endpointSuffix: string
+				}
+				azureKustoResource: string,
+				powerBiResource: string,
+				scopes: string,
+				portalEndpoint: string
+			}
+		}
+	}
 }
 
 export interface Subscription {
