+++ extensions/azurecore/src/extension.ts
@@ -293,25 +293,11 @@ async function onDidChangeConfiguration(e: vscode.ConfigurationChangeEvent): Pro
 		if (vscode.workspace.getConfiguration(Constants.AzureSection).get('authenticationLibrary') === 'ADAL') {
 			void vscode.window.showInformationMessage(loc.deprecatedOption);
 		}
-		await displayReloadAds('authenticationLibrary');
+		await utils.displayReloadAds('authenticationLibrary');
 	}
 }
 
 function updatePiiLoggingLevel(): void {
 	const piiLogging: boolean = vscode.workspace.getConfiguration(Constants.AzureSection).get('piiLogging', false);
 	Logger.piiLogging = piiLogging;
-}
-
-// Display notification with button to reload
-// return true if button clicked
-// return false if button not clicked
-export async function displayReloadAds(sectionName: string): Promise<boolean> {
-	const result = await vscode.window.showInformationMessage(loc.reloadPrompt(sectionName), loc.reloadChoice);
-	if (result === loc.reloadChoice) {
-		await vscode.commands.executeCommand('workbench.action.reloadWindow');
-		return true;
-	} else {
-		return false;
-	}
-
 }
